NutriCUT/admin/requirements.txt
Django==3.2
djangorestframework==3.12.2
djangorestframework-simplejwt==5.3.1
pyjwt==2.8.0
mysqlclient==2.0.1
django-mysql==3.9
django-cors-headers==3.5.0
pika==1.3.2
